refactor: game_scene.js
