+++ front/srcs/game/map_image_generator.js
