+++ front/srcs/game/physics.js
@@ -1,5 +1,5 @@
-import  PhysicsEntity, { isCircleCollideRect, isRectCollideRect } from "@/game/physicsEntity"
-import { isEqualF } from "@/game/physicsUtils";
+import  PhysicsEntity, { isCircleCollideRect, isRectCollideRect } from "@/game/physics_entity"
+import { isEqualF } from "@/game/physics_utils";
 
 export default class Physics {
 
@@ -82,14 +82,14 @@ export default class Physics {
     }
     const state = {
       accel: obj.acceleration,
-      velocity: obj.veolocity,
+      velocity: obj.velocity,
       position: obj.position
     };
     const res = setCallback({...state});
     if (res.accel)
       obj.acceleration = res.accel;
     if (res.velocity)
-      obj.veolocity = res.velocity;
+      obj.velocity = res.velocity;
     if (res.position)
       obj.position = res.position;
   }
@@ -115,7 +115,7 @@ export default class Physics {
             x: obj.midX,
             y: obj.midY
           },
-          veolocity: {...obj.veolocity}
+          velocity: {...obj.velocity}
         };
       });
     return states;
@@ -128,7 +128,7 @@ export default class Physics {
       throw "Fail to get object for " + objId;
     return ({
       position: {...obj.position},
-      veolocity: {...obj.veolocity}
+      velocity: {...obj.velocity}
     });
   }
 
@@ -154,7 +154,7 @@ export default class Physics {
   #updateVelocities(elapsedTime) {
     for (let id of this.#movableObjects) {
       const obj = this.#allObjects[id];
-      const start = {...obj.veolocity};
+      const start = {...obj.velocity};
       const accel = obj.acceleration;
       if (isEqualF(accel.x, 0))
         accel.x = 0;
@@ -168,7 +168,7 @@ export default class Physics {
         after.x = 0;
       if (isEqualF(after.y, 0))
         after.y = 0;
-      obj.veolocity = after;
+      obj.velocity = after;
     }
     return this;
   }
@@ -178,7 +178,7 @@ export default class Physics {
     for (let id of this.#movableObjects) {
       const obj = this.#allObjects[id];
       const start = {...obj.position};
-      const vel = obj.veolocity;
+      const vel = obj.velocity;
       const after = {
         x: start.x + vel.x * elapsedTime,
         y: start.y + vel.y * elapsedTime
@@ -240,14 +240,14 @@ export default class Physics {
         return ;
       }
 
-      const collisionEpsilon = (Math.abs(collider.veolocity.x) + Math.abs(collider.veolocity.y)) * 0.05;
+      const collisionEpsilon = (Math.abs(collider.velocity.x) + Math.abs(collider.velocity.y)) * 0.05;
       if (collideAxes.x) {
-        if (collider.veolocity.x > 0 && 
+        if (collider.velocity.x > 0 && 
           collider.midX < collidee.midX &&
           Math.abs(collider.right - collidee.left) < collisionEpsilon) {
           collider.position.x = collidee.left - collider.width; 
         }
-        else if (collider.veolocity.x < 0 && 
+        else if (collider.velocity.x < 0 && 
           collider.midX > collidee.midX &&
           Math.abs(collider.left - collidee.right) < collisionEpsilon) {
           collider.position.x = collidee.right;
@@ -257,12 +257,12 @@ export default class Physics {
         }
       }
       if (collideAxes.y) {
-        if (collider.veolocity.y > 0 && 
+        if (collider.velocity.y > 0 && 
           collider.midY < collidee.midY &&
           Math.abs(collider.top - collidee.bottom) < collisionEpsilon) {
           collider.position.y = collidee.bottom - collider.height;
         }
-        else if (collider.veolocity.y < 0 &&
+        else if (collider.velocity.y < 0 &&
           collider.midY > collidee.midY &&
           Math.abs(collider.bottom - collidee.top) < collisionEpsilon) {
           collider.position.y = collidee.top;
@@ -275,26 +275,26 @@ export default class Physics {
       if (collider.isDynamic)  {
         // TODO: acceleration?
         if (collideAxes.x) { 
-          collider.veolocity.x *= -1;
+          collider.velocity.x *= -1;
         }
         if (collideAxes.y)
-          collider.veolocity.y *= -1;
+          collider.velocity.y *= -1;
 
         if (collideAxes.x && collideAxes.y && collider.isShape("CIRCLE")) {
           let randomness = (Math.random() - 0.5) * 0.2 ;
-          if (collider.veolocity.x > 0) {
-            collider.veolocity.x += randomness;
-            collider.veolocity.y -= randomness;
+          if (collider.velocity.x > 0) {
+            collider.velocity.x += randomness;
+            collider.velocity.y -= randomness;
           }
           else {
-            collider.veolocity.x -= randomness;
-            collider.veolocity.y += randomness;
+            collider.velocity.x -= randomness;
+            collider.velocity.y += randomness;
           }
         }
       }
       else {
         collider.acceleration = { x: 0, y: 0 };
-        collider.veolocity = { x: 0, y:0 };
+        collider.velocity = { x: 0, y:0 };
       }
     }
