+++ front/srcs/utils/math_util.js
