+++ front/srcs/utils/three_util.js
@@ -1,0 +1,10 @@
+import * as THREE from "three";
+
+export function resizeTexture({texture, x, y}) {
+  texture.repeat.x = x;
+  texture.repeat.y = y;
+  texture.wrapS = THREE.RepeatWrapping;
+  texture.wrapT = THREE.RepeatWrapping;
+  texture.magFilter = THREE.LinearFilter; 
+  texture.minFilter = THREE.NearestMipMapLinearFilter;
+}
