+++ front/srcs/utils/type_util.js
