+++ front/srcs/views/components/user_label.html
@@ -2,10 +2,10 @@
   <style>
     .label-container {
       width: 80%;
-      height: 80%;
+      height: 90%;
       display: flex;
       flex-direction: column;
-      justify-content: center;
+      justify-content: flex-start;
       align-items: center;
       position:relative;
       background:#FFF455;
@@ -50,19 +50,22 @@
       bottom:0px;
       left:10%;
     }
-    h1 {
+    h2 {
       position:relative;
+      flex-basis: 20%;
     }
     ul {
       align-self: flex-start;
     }
     li {
-      font-size: 15px;
+      list-style: none;
+      font-size: 24px;
+      font-weight: bold;
     }
 
   </style>
   <div class="label-container">
-    <h1 id="label-user-name">{{name}}</h1>
+    <h2 id="label-user-name">{{name}}</h1>
     <ul data-condition="texts" data-for="text in texts">
       <li>{{text}}</li>
     </ul>
