+++ front/srcs/views/home/home_view.html
@@ -1,5 +1,6 @@
 <home-view class="d-flex flex-column justify-content-center view">
-<div class="mb-auto">
+  <color-picker data-component="ColorPicker"></color-picker>
+<!--<div class="mb-auto">
     <nav-bar data-component="NavBar"></nav-bar>
   </div>
   <div class="row align-items-center justify-content-center" style="flex-grow: 1;">
@@ -7,4 +8,5 @@
       <a href="/mode" data-link><img src="/srcs/assets/start.png" alt="game_start" style="height: 50vh; width:auto;"></a>
     </div>
   </div>
+-->
 </home-view>
