+++ front/srcs/views/home/home_view.js
@@ -1,7 +1,6 @@
 import View from "@/lib/view";
 import ObservableObject from "@/lib/observable_object";
 import User from "@/data/user";
-import MapImageGenarator from "@/game/mapImageGenerator";
 
 export default class HomeView extends View {
